Split large relational test modules
