+++ crates/truthdb-core/src/engine/relational/tests/query/mod.rs
@@ -1,0 +1,4 @@
+use super::*;
+
+mod planning_and_spilling;
+mod sql_behavior;
